Extract the error interfaces from the Course Modal and update imports
This commit extracts the error interfaces that were defined in the Course Modal into a separate file so that the other components are not importing the interfaces from the CourseModal component.
