+++ src/client/classes/index.ts
@@ -1,1 +1,2 @@
 export * from './AppMessage';
+export * from './Validation';
